+++ bonheur/Sources/Models/NetworkResult.swift
@@ -1,0 +1,8 @@
+//
+//  NetworkResult.swift
+//  bonheur
+//
+//  Created by 이정연 on 2023/01/09.
+//
+
+import Foundation
